validate categry only admin can post

+++ src/main/java/com/kshrd/wearekhmer/Category/controller/CategoryController.java
@@ -4,10 +4,14 @@ package com.kshrd.wearekhmer.Category.controller;
 import com.kshrd.wearekhmer.Category.model.Category;
 import com.kshrd.wearekhmer.Category.model.CategoryRequestDTO;
 import com.kshrd.wearekhmer.Category.service.CategoryService;
+import com.kshrd.wearekhmer.exception.CustomRuntimeException;
 import com.kshrd.wearekhmer.requestRequest.GenericResponse;
-import jakarta.annotation.PostConstruct;
+import com.kshrd.wearekhmer.utils.validation.WeAreKhmerValidation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,10 +19,13 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/category")
+@SecurityRequirement(name = "bearerAuth")
 @AllArgsConstructor
-public class CategoryController implements CategoryControllerInterface{
+public class CategoryController implements CategoryControllerInterface {
 
     private final CategoryService categoryService;
+
+    private WeAreKhmerValidation weAreKhmerValidation;
 
 
     @Override
@@ -29,7 +36,7 @@ public class CategoryController implements CategoryControllerInterface{
         GenericResponse genericResponse;
         try {
             List<Category> categories = categoryService.getAllCategories();
-             genericResponse =
+            genericResponse =
                     GenericResponse.builder()
                             .status("200")
                             .message("success")
@@ -78,7 +85,14 @@ public class CategoryController implements CategoryControllerInterface{
     @PostMapping
 
     @Override
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @SecurityRequirement(name = "bearerAuth")
+    @Operation(summary = "(Only administrator can create this category.)")
     public ResponseEntity<?> insertCategory(CategoryRequestDTO category) {
+        if (!weAreKhmerValidation.isAdmin()) {
+            throw new CustomRuntimeException("You are not administrator.");
+        }
+
         Category category1 = Category.builder()
                 .categoryName(category.getCategoryName())
                 .build();
@@ -107,7 +121,12 @@ public class CategoryController implements CategoryControllerInterface{
 
     @Override
     @PutMapping
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @Operation(summary = "(Only administrator can update this category.)")
     public ResponseEntity<?> updateCategory(Category category) {
+        if (!weAreKhmerValidation.isAdmin()) {
+            throw new CustomRuntimeException("You are not administrator.");
+        }
         GenericResponse genericResponse;
         try {
             Category category2 = categoryService.updateCategory(category);
@@ -133,7 +152,12 @@ public class CategoryController implements CategoryControllerInterface{
 
     @Override
     @DeleteMapping
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @Operation(summary = "(Only administrator can delete this category.)")
     public ResponseEntity<?> deleteCategory(String categoryId) {
+        if (!weAreKhmerValidation.isAdmin()) {
+            throw new CustomRuntimeException("You are not administrator.");
+        }
         GenericResponse genericResponse;
         try {
             Category category2 = categoryService.deleteCategory(categoryId);
